move some modules around
